+++ README.md
@@ -1,23 +1,23 @@
 # Counter Smart Contract
 
-This is an example of a simple smart contract, the Counter.
+This is an example of a simple smart contract.
 
 ## Usage
 
 ### Set up the Rust toolchain
-You need the Rust toolchain to develop smart contracts.
+
 ```bash
 make prepare
 ```
 
 ### Compile smart contracts
-Compile WASM files that will be used later.
+
 ```bash
-make build-contracts
+make build-contract
 ```
 
 ### Run tests
-You can run integration tests with this command:
+
 ```bash
 make test
 ```
