+++ contract/Cargo.toml
@@ -1,7 +1,6 @@
 [package]
 name = "counter-define"
 version = "1.0.0"
-authors = ["Maciej Zielinski <maciej@casperlabs.io>"]
 edition = "2021"
 
 [dependencies]
@@ -14,7 +13,3 @@ path = "src/main.rs"
 bench = false
 doctest = false
 test = false
-
-[profile.release]
-codegen-units = 1
-lto = true
