+++ counter-call/Cargo.toml
@@ -14,7 +14,3 @@ path = "src/main.rs"
 bench = false
 doctest = false
 test = false
-
-[profile.release]
-codegen-units = 1
-lto = true
