+++ tests/Cargo.toml
@@ -1,10 +1,9 @@
 [package]
 name = "tests"
 version = "1.0.0"
-authors = ["Maciej Zielinski <maciek.s.zielinski@gmail.com>"]
 edition = "2021"
 
-[dev-dependencies]
+[dependencies]
 casper-engine-test-support = { version = "2.2.0", features = ["test-support"] }
 casper-execution-engine = "2.0.0"
 casper-types = "1.5.0"
